reorganize report module into submodules
