+++ crates/cli/src/commands/report/chart/mod.rs
@@ -1,8 +1,10 @@
+mod chart_id;
 mod gas_per_block;
 mod heatmap;
 mod time_to_inclusion;
 mod tx_gas_used;
 
+pub use chart_id::ReportChartId;
 pub use gas_per_block::GasPerBlockChart;
 pub use heatmap::HeatMapChart;
 pub use time_to_inclusion::TimeToInclusionChart;
