+++ crates/cli/src/commands/report/mod.rs
@@ -1,112 +1,25 @@
+mod block_trace;
+mod cache;
 mod chart;
 mod gen_html;
 mod util;
 
 use crate::util::{data_dir, write_run_txs};
-use alloy::providers::ext::DebugApi;
-use alloy::rpc::types::Block;
-use alloy::{
-    providers::{Provider, ProviderBuilder},
-    rpc::types::{
-        trace::geth::{
-            GethDebugBuiltInTracerType, GethDebugTracerConfig, GethDebugTracerType,
-            GethDebugTracingOptions, GethDefaultTracingOptions, GethTrace,
-        },
-        TransactionReceipt,
-    },
-    transports::http::reqwest::Url,
-};
+use alloy::{providers::ProviderBuilder, transports::http::reqwest::Url};
+use block_trace::get_block_trace_data;
+use cache::CacheFile;
+use chart::ReportChartId;
 use chart::{GasPerBlockChart, HeatMapChart, TimeToInclusionChart, TxGasUsedChart};
-use contender_core::{
-    db::{DbOps, RunTx},
-    generator::types::EthProvider,
-};
+use contender_core::db::{DbOps, RunTx};
 use csv::WriterBuilder;
 use gen_html::build_html_report;
-use serde::{Deserialize, Serialize};
 use std::str::FromStr;
-
-static CACHE_FILENAME: &str = "debug_trace.json";
-
-#[derive(Serialize, Deserialize)]
-struct CacheFile {
-    traces: Vec<TxTraceReceipt>,
-    blocks: Vec<Block>,
-}
-
-impl CacheFile {
-    fn new(traces: Vec<TxTraceReceipt>, blocks: Vec<Block>) -> Self {
-        Self { traces, blocks }
-    }
-
-    /// Returns the fully-qualified path to the cache file.
-    fn cache_path() -> Result<String, Box<dyn std::error::Error>> {
-        Ok(format!("{}/{}", data_dir()?, CACHE_FILENAME))
-    }
-
-    fn load() -> Result<Self, Box<dyn std::error::Error>> {
-        let file = std::fs::File::open(CacheFile::cache_path()?)?;
-        let cache_data: CacheFile = serde_json::from_reader(file)?;
-        Ok(cache_data)
-    }
-
-    fn save(&self) -> Result<(), Box<dyn std::error::Error>> {
-        let file = std::fs::File::create(CacheFile::cache_path()?)?;
-        serde_json::to_writer(file, self)?;
-        Ok(())
-    }
-}
 
 /// Returns the fully-qualified path to the report directory.
 fn report_dir() -> Result<String, Box<dyn std::error::Error>> {
     let path = format!("{}/reports", data_dir()?);
     std::fs::create_dir_all(&path)?;
     Ok(path)
-}
-
-enum ReportChartId {
-    Heatmap,
-    GasPerBlock,
-    TimeToInclusion,
-    TxGasUsed,
-}
-
-impl std::fmt::Display for ReportChartId {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        let s = match self {
-            ReportChartId::Heatmap => "heatmap",
-            ReportChartId::GasPerBlock => "gas_per_block",
-            ReportChartId::TimeToInclusion => "time_to_inclusion",
-            ReportChartId::TxGasUsed => "tx_gas_used",
-        };
-        write!(f, "{}", s)
-    }
-}
-
-impl ReportChartId {
-    fn filename(
-        &self,
-        start_run_id: u64,
-        end_run_id: u64,
-    ) -> Result<String, Box<dyn std::error::Error>> {
-        Ok(format!(
-            "{}/{}_run-{}-{}.png",
-            report_dir()?,
-            self,
-            start_run_id,
-            end_run_id
-        ))
-    }
-
-    fn proper_name(&self) -> String {
-        match self {
-            ReportChartId::Heatmap => "Storage Slot Heatmap",
-            ReportChartId::GasPerBlock => "Gas Per Block",
-            ReportChartId::TimeToInclusion => "Time To Inclusion",
-            ReportChartId::TxGasUsed => "Tx Gas Used",
-        }
-        .to_string()
-    }
 }
 
 pub async fn report(
@@ -175,89 +88,6 @@ pub async fn report(
     webbrowser::open(&report_path)?;
 
     Ok(())
-}
-
-#[derive(Clone, Debug, Deserialize, Serialize)]
-struct TxTraceReceipt {
-    trace: GethTrace,
-    receipt: TransactionReceipt,
-}
-
-impl TxTraceReceipt {
-    fn new(trace: GethTrace, receipt: TransactionReceipt) -> Self {
-        Self { trace, receipt }
-    }
-}
-
-async fn get_block_trace_data(
-    txs: &[RunTx],
-    rpc_client: &EthProvider,
-) -> Result<(Vec<TxTraceReceipt>, Vec<Block>), Box<dyn std::error::Error>> {
-    if std::env::var("DEBUG_USEFILE").is_ok() {
-        println!("DEBUG_USEFILE detected: using cached data");
-        // load trace data from file
-        let cache_data = CacheFile::load()?;
-        return Ok((cache_data.traces, cache_data.blocks));
-    }
-
-    // find block range of txs
-    let (min_block, max_block) = txs.iter().fold((u64::MAX, 0), |(min, max), tx| {
-        (min.min(tx.block_number), max.max(tx.block_number))
-    });
-
-    // pad block range on each side
-    let block_padding = 3;
-    let min_block = min_block - block_padding;
-    let max_block = max_block + block_padding;
-
-    // get block data
-    let mut all_blocks = vec![];
-    for block_num in min_block..=max_block {
-        let block = rpc_client
-            .get_block_by_number(block_num.into(), true)
-            .await?;
-        if let Some(block) = block {
-            println!("read block {}", block.header.number);
-            all_blocks.push(block);
-        }
-    }
-
-    // get tx traces for all txs in all_blocks
-    let mut all_traces = vec![];
-    for block in &all_blocks {
-        for tx_hash in block.transactions.hashes() {
-            println!("tracing tx {:?}", tx_hash);
-            let trace = rpc_client
-                .debug_trace_transaction(
-                    tx_hash,
-                    GethDebugTracingOptions {
-                        config: GethDefaultTracingOptions::default(),
-                        tracer: Some(GethDebugTracerType::BuiltInTracer(
-                            GethDebugBuiltInTracerType::PreStateTracer,
-                        )),
-                        tracer_config: GethDebugTracerConfig::default(),
-                        timeout: None,
-                    },
-                )
-                .await?;
-
-            // receipt might fail if we target a non-ETH chain
-            // so if it does fail, we just ignore it
-            let receipt = rpc_client.get_transaction_receipt(tx_hash).await;
-            if let Ok(receipt) = receipt {
-                if let Some(receipt) = receipt {
-                    println!("got receipt for tx {:?}", tx_hash);
-                    all_traces.push(TxTraceReceipt::new(trace, receipt));
-                } else {
-                    println!("no receipt for tx {:?}", tx_hash);
-                }
-            } else {
-                println!("ignored receipt for tx {:?} (failed to decode)", tx_hash);
-            }
-        }
-    }
-
-    Ok((all_traces, all_blocks))
 }
 
 /// Saves RunTxs to `{data_dir}/reports/{id}.csv`.
